double coordinates

+++ app/src/main/java/com/navapp/data/AddressModel.java
@@ -18,11 +18,11 @@ public class AddressModel {
     private String rawAddress;
 
     @ColumnInfo(name = "longitude")
-    private long longitude;
+    private double longitude;
     @ColumnInfo(name = "latitude")
-    private long latitude;
+    private double latitude;
 
-    public AddressModel(@NonNull String rawAddress, long longitude, long latitude) {
+    public AddressModel(@NonNull String rawAddress, double longitude, double latitude) {
         this.rawAddress = rawAddress;
         this.longitude = longitude;
         this.latitude = latitude;
@@ -43,17 +43,17 @@ public class AddressModel {
         this.rawAddress = rawAddress;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
